Extract bounding box
This function is specifically designed to manipulate files with .vtp extension
The information contained in these files will serve as basis to determine the bounding box of the standard point cloud.
The pose of the captured object will be read from the files to apply the transformation to the standard bounding box in order to obtain the current bounding box position.

signed-by: mariem.khlifi@tum.de